+++ lib/feature/shared/presentation/ui/admin/admin_user_details.dart
@@ -28,22 +28,29 @@ class AdminUserDetails extends StatefulWidget {
   const AdminUserDetails({super.key, this.user});
 
   @override
-  State<AdminUserDetails> createState() => _AdminUserDetailsState();
+  _AdminUserDetailsState createState() => _AdminUserDetailsState();
 }
 
 class _AdminUserDetailsState extends State<AdminUserDetails> {
+  // Controllers for handling user input
   late final TextEditingController emailController;
   late final TextEditingController passController;
   late final TextEditingController phoneNumController;
   late final TextEditingController usernameController;
   late final TextEditingController employeeNumberController;
+
+  // DTO for storing updated user information
   final UserDTO updatedUser = UserDTO();
+
+  // Selected indices for role and gender dropdowns
   late int? selectedRoleIndex;
   late int? selectedGenderIndex;
 
   @override
   void initState() {
     super.initState();
+
+    // Initialize controllers with user data
     final user = widget.user!;
     emailController = TextEditingController(text: user.email);
     passController = TextEditingController(text: user.passwordHash);
@@ -51,194 +58,198 @@ class _AdminUserDetailsState extends State<AdminUserDetails> {
     usernameController = TextEditingController(text: user.username);
     employeeNumberController = TextEditingController(text: user.employeeNumber);
 
+    // Set initial selected indices
     selectedRoleIndex = user.role!.id! - 1;
     selectedGenderIndex = user.gender!.id! - 1;
   }
 
   @override
   Widget build(BuildContext context) {
+    // Check if the screen size is mobile
     final isMobile = Responsive.isMobile(context);
-    final List<String> fields = (widget.user!.role!.roleName != "user")
+
+    // Determine which fields to display based on user role
+    final fields = (widget.user!.role!.roleName != "user")
         ? ["username", "password", "phone num", "email", "employee number"]
         : ["username", "password", "phone num", "email"];
+
     return Scaffold(
       body: SafeArea(
         child: SingleChildScrollView(
-          child: Padding(
-            padding: const EdgeInsets.all(16.0),
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                Row(
-                  children: [
-                    IconButton(
-                      padding: EdgeInsets.zero,
-                      onPressed: () {
-                        Navigator.pop(context);
-                      },
-                      icon: Icon(
-                        Icons.arrow_back,
-                        color: AppColors.darkBrown,
-                        size: 30,
-                      ),
+          padding: const EdgeInsets.all(16.0),
+          child: Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: [
+              // Header and back button
+              Row(
+                children: [
+                  IconButton(
+                    padding: EdgeInsets.zero,
+                    onPressed: () {
+                      Navigator.pop(context);
+                    },
+                    icon: Icon(
+                      Icons.arrow_back,
+                      color: AppColors.darkBrown,
+                      size: 30,
                     ),
-                    Expanded(
-                      child: Row(
-                        mainAxisAlignment: MainAxisAlignment.start,
-                        children: [
-                          Expanded(
-                            child: HeaderText(
-                              textSize: isMobile ? 35 : 45,
-                              title: 'user details',
-                            ),
+                  ),
+                  Expanded(
+                    child: Row(
+                      mainAxisAlignment: MainAxisAlignment.start,
+                      children: [
+                        Expanded(
+                          child: HeaderText(
+                            textSize: isMobile ? 35 : 45,
+                            title: 'user details',
                           ),
-                        ],
-                      ),
-                    ),
-                  ],
-                ),
-                Row(
-                  mainAxisAlignment: MainAxisAlignment.end,
-                  children: [
-                    Icon(Icons.group, size: 500, color: AppColors.darkBrown),
-                    Expanded(
-                      child: Column(
-                        mainAxisAlignment: MainAxisAlignment.spaceAround,
-                        children: [
-                          for (var field in fields)
-                            BasicTextField(
-                              title: field,
-                              controller: _getControllerForField(field),
-                              isEnabled: field != "password",
-                            ),
-                          Row(
-                            children: [
-                              Expanded(
-                                child: BlocProvider<RemoteRoleBloc>(
-                                  create: (context) =>
-                                      service()..add(const GetRoles()),
-                                  child: BlocBuilder<RemoteRoleBloc,
-                                      RemoteRoleState>(
-                                    builder: (context, state) {
-                                      switch (state) {
-                                        case RemoteRoleLoading():
-                                          return const SizedBox();
-                                        case RemoteRoleDone():
-                                          return BasicDropdown(
-                                            listTitle: "role",
-                                            dropdownData: state.roles!
-                                                .map((e) => e.roleName!)
-                                                .toList(),
-                                            selectedIndex: selectedRoleIndex!,
-                                            onIndexChanged: (value) {
-                                              selectedRoleIndex = value;
-                                            },
-                                          );
-                                        case RemoteRoleError():
-                                          return const Text("error");
-                                      }
-                                      return const SizedBox();
-                                    },
-                                  ),
-                                ),
-                              ),
-                              Expanded(
-                                child: BlocProvider<RemoteGendersBloc>(
-                                  create: (context) =>
-                                      service()..add(const GetGenders()),
-                                  child: BlocBuilder<RemoteGendersBloc,
-                                      RemoteGenderState>(
-                                    builder: (_, state) {
-                                      switch (state) {
-                                        case RemoteGenderLoading():
-                                          return const SizedBox();
-                                        case RemoteGenderDone():
-                                          return BasicDropdown(
-                                            listTitle: "gender",
-                                            dropdownData: state.genders!
-                                                .map((e) => e.nameCategory!)
-                                                .toList()
-                                                .reversed
-                                                .toList(),
-                                            selectedIndex: selectedGenderIndex!,
-                                            onIndexChanged: (value) {
-                                              selectedGenderIndex = value;
-                                            },
-                                          );
-                                        case RemoteGenderError():
-                                          return const Text("error");
-                                      }
-                                      return const SizedBox();
-                                    },
-                                  ),
-                                ),
-                              ),
-                            ],
-                          ),
-                        ],
-                      ),
-                    )
-                  ],
-                ),
-                if (widget.user!.role!.roleName == "user") ...[
-                  HeaderText(
-                    textSize: isMobile ? 35 : 45,
-                    title: 'user orders',
-                  ),
-                  Padding(
-                    padding: const EdgeInsets.all(8.0),
-                    child: BasicContainer(
-                      child: Column(
-                        children: [
-                          UserOrdersTable(user: widget.user!),
-                        ],
-                      ),
-                    ),
-                  ),
-                  HeaderText(
-                    textSize: isMobile ? 35 : 45,
-                    title: 'user payment info',
-                  ),
-                  const Padding(
-                    padding: EdgeInsets.all(8.0),
-                    child: BasicContainer(
-                      child: Column(
-                        children: [
-                          BasicText(title: 'user payment info'),
-                        ],
-                      ),
-                    ),
-                  ),
-                  HeaderText(
-                    textSize: isMobile ? 35 : 45,
-                    title: 'user addresses',
-                  ),
-                  const Padding(
-                    padding: EdgeInsets.all(8.0),
-                    child: BasicContainer(
-                      child: Column(
-                        children: [
-                          BasicText(title: 'user addresses'),
-                        ],
-                      ),
+                        ),
+                      ],
                     ),
                   ),
                 ],
-                Row(
-                  mainAxisAlignment: MainAxisAlignment.end,
-                  children: [
-                    _buildButton("Edit", AppColors.darkBrown, _editUserInfo),
-                    _buildButton("Delete", AppColors.red, _deleteUser),
-                  ],
+              ),
+              // User input fields and dropdowns
+              Row(
+                mainAxisAlignment: MainAxisAlignment.end,
+                children: [
+                  Icon(Icons.group, size: 500, color: AppColors.darkBrown),
+                  Expanded(
+                    child: Column(
+                      mainAxisAlignment: MainAxisAlignment.spaceAround,
+                      children: [
+                        // Text fields for user input
+                        for (var field in fields)
+                          BasicTextField(
+                            title: field,
+                            controller: _getControllerForField(field),
+                            isEnabled: field != "password",
+                          ),
+                        // Dropdowns for role and gender
+                        Row(
+                          children: [
+                            Expanded(
+                              child: BlocProvider<RemoteRoleBloc>(
+                                create: (context) =>
+                                    service()..add(const GetRoles()),
+                                child: BlocBuilder<RemoteRoleBloc,
+                                    RemoteRoleState>(
+                                  builder: (context, state) {
+                                    if (state is RemoteRoleDone) {
+                                      return BasicDropdown(
+                                        listTitle: "role",
+                                        dropdownData: state.roles!
+                                            .map((e) => e.roleName!)
+                                            .toList(),
+                                        selectedIndex: selectedRoleIndex!,
+                                        onIndexChanged: (value) {
+                                          setState(() {
+                                            selectedRoleIndex = value;
+                                          });
+                                        },
+                                      );
+                                    }
+                                    return const SizedBox();
+                                  },
+                                ),
+                              ),
+                            ),
+                            Expanded(
+                              child: BlocProvider<RemoteGendersBloc>(
+                                create: (context) =>
+                                    service()..add(const GetGenders()),
+                                child: BlocBuilder<RemoteGendersBloc,
+                                    RemoteGenderState>(
+                                  builder: (_, state) {
+                                    if (state is RemoteGenderDone) {
+                                      return BasicDropdown(
+                                        listTitle: "gender",
+                                        dropdownData: state.genders!
+                                            .map((e) => e.nameCategory!)
+                                            .toList()
+                                            .reversed
+                                            .toList(),
+                                        selectedIndex: selectedGenderIndex!,
+                                        onIndexChanged: (value) {
+                                          setState(() {
+                                            selectedGenderIndex = value;
+                                          });
+                                        },
+                                      );
+                                    }
+                                    return const SizedBox();
+                                  },
+                                ),
+                              ),
+                            ),
+                          ],
+                        ),
+                      ],
+                    ),
+                  )
+                ],
+              ),
+              // Additional sections if user role is "user"
+              if (widget.user!.role!.roleName == "user") ...[
+                HeaderText(
+                  textSize: isMobile ? 35 : 45,
+                  title: 'user orders',
+                ),
+                Padding(
+                  padding: const EdgeInsets.all(8.0),
+                  child: BasicContainer(
+                    child: Column(
+                      children: [
+                        UserOrdersTable(user: widget.user!),
+                      ],
+                    ),
+                  ),
+                ),
+                HeaderText(
+                  textSize: isMobile ? 35 : 45,
+                  title: 'user payment info',
+                ),
+                const Padding(
+                  padding: EdgeInsets.all(8.0),
+                  child: BasicContainer(
+                    child: Column(
+                      children: [
+                        BasicText(title: 'user payment info'),
+                      ],
+                    ),
+                  ),
+                ),
+                HeaderText(
+                  textSize: isMobile ? 35 : 45,
+                  title: 'user addresses',
+                ),
+                const Padding(
+                  padding: EdgeInsets.all(8.0),
+                  child: BasicContainer(
+                    child: Column(
+                      children: [
+                        BasicText(title: 'user addresses'),
+                      ],
+                    ),
+                  ),
                 ),
               ],
-            ),
+              // Buttons for editing and deleting user
+              Row(
+                mainAxisAlignment: MainAxisAlignment.end,
+                children: [
+                  _buildButton("Edit", AppColors.darkBrown, _editUserInfo),
+                  _buildButton("Delete", AppColors.red, _deleteUser),
+                ],
+              ),
+            ],
           ),
         ),
       ),
     );
   }
 
+  // Function to handle updating user information
   void _editUserInfo() async {
     updatedUser.id = widget.user!.id!;
     updatedUser.email = emailController.text;
@@ -248,6 +259,7 @@ class _AdminUserDetailsState extends State<AdminUserDetails> {
     updatedUser.employeeNumber = employeeNumberController.text;
     updatedUser.gender = selectedGenderIndex! + 1;
     updatedUser.role = selectedRoleIndex! + 1;
+
     service<RemoteUserBloc>().add(UpdateUser(user: updatedUser));
     await Future.delayed(const Duration(seconds: 1));
     router.pop();
@@ -256,16 +268,22 @@ class _AdminUserDetailsState extends State<AdminUserDetails> {
     );
   }
 
+  // Function to handle deleting user
   void _deleteUser() async {
-    //TODO: Add logic for deleting user
-    service<RemoteUserBloc>().add(DeleteUser(id: widget.user!.id!));
-    await Future.delayed(const Duration(seconds: 1));
-    router.pop();
-    router.push(
-      Pages.adminAllUsers.screenPath,
-    );
+    if (widget.user!.role!.roleName! != "user") {
+      service<RemoteUserBloc>().add(DeleteUser(id: widget.user!.id!));
+      await Future.delayed(const Duration(seconds: 1));
+      router.pop();
+      router.push(
+        Pages.adminAllUsers.screenPath,
+      );
+    } else {
+      //TODO: handle case when user wants to delete buyer (that's not allowed)
+      router.pop();
+    }
   }
 
+  // Function to get the controller for a specific input field
   TextEditingController _getControllerForField(String field) {
     switch (field) {
       case "username":
@@ -274,17 +292,16 @@ class _AdminUserDetailsState extends State<AdminUserDetails> {
         return passController;
       case "phone num":
         return phoneNumController;
-
       case "email":
         return emailController;
       case "employee number":
         return employeeNumberController;
-
       default:
         throw Exception("Invalid field: $field");
     }
   }
 
+  // Widget for building custom buttons
   Widget _buildButton(
       String title, Color backgroundColor, VoidCallback onPressed) {
     return Padding(
